Volcanic Blast VFX
A more accurate local space Volcanic Blast VFX has been created through utilising meshes. This was done through utilising size over lifetime for the initial animation, and an unlit shader graph for everything else. The graph's core works use a series of properties in order to create the "core colour", a coloured simple noise which helps make the blast look more authentic and animated. Alongside this, a UV channel hooked to the alpha gives the ability for the system to give control of the alpha and fade out. 

In order to not take so much screen space, the blast has been appropriately scaled down and currently a ripple/after effect is in the works to add to the animation, as well as the properties of the spell itself in coding.