Volcanic Blast Balancing and Visual Progress
The previous Blast visuals have been attached to the existing enveloping circle that dictates damage for the spell. Initial balancing efforts decreased the amount of damage drastically as it was thought to be a multi-hit spell however this was dropped once the contrary was found. A Vector3.Distance method was also planned out, before being dropped as likewise with the balancing, it was implemented with the belief of the spell being multi-hit. 

As the appearance of the circle did not match too well with the visuals and the width of the mesh itself was too thick, a new Cylinder to base it off was created in Blender and ported over alongside the replication of the blast material for more appropriate use with the cylinder. In terms of polishing the spell up, the blast visual simply needs to become more appealing with the use of it in the blast spell, as it now lasts 1 second instead of 4 and thus looks a little sudden.